feat(surveillance-views): add GET/PUT REST resource for surveillance-view-config
Implements SurveillanceViewConfigRestService at /api/v2/surveillance-view-config
with JSON GET/PUT endpoints backed by JaxbUtils marshaling of surveillance-views.xml.
Fixes assertEquals ambiguity in the test due to View.getRefreshSeconds() returning Integer.

+++ opennms-webapp-rest/src/test/java/org/opennms/web/rest/v2/SurveillanceViewConfigRestServiceTest.java
@@ -118,7 +118,7 @@ public class SurveillanceViewConfigRestServiceTest {
         assertEquals(1, roundTripped.getViews().size());
         View rt = roundTripped.getViews().get(0);
         assertEquals("default", rt.getName());
-        assertEquals(300, rt.getRefreshSeconds());
+        assertEquals(300, (int) rt.getRefreshSeconds());
         assertEquals(1, rt.getRows().size());
         assertEquals("Switches", rt.getRows().get(0).getLabel());
         assertEquals("Switches", rt.getRows().get(0).getCategories().get(0).getName());
